use new `custom` fn. mv fn to src/tcp_server.rs
