+++ Cargo.toml
@@ -44,6 +44,7 @@ parking_lot = "0.12"
 radix_trie = "0.2"
 rustc-hash = "1.1.0"
 simplelog = "0.12.0"
+serde_json = { version = "1", features = ["alloc"], default_features = false, optional = true }
 
 # kanata-keyberon = "0.160.1"
 # kanata-parser = "0.160.1"
@@ -80,7 +81,7 @@ kanata-interception = { version = "0.2.0", optional = true }
 [features]
 default = ["tcp_server"]
 perf_logging = []
-tcp_server = []
+tcp_server = ["serde_json"]
 win_sendinput_send_scancodes = []
 cmd = ["kanata-parser/cmd"]
 interception_driver = ["kanata-interception", "kanata-parser/interception_driver"]
